+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## master
 
 * Fixed Docker image outdated ruby version by updating it to 2.7. [@axl89](https://github.com/axl89) [#1383](https://github.com/danger/danger/issues/1383)
+* Added support for [Appcircle](https://appcircle.io/) - [@tosbaha](https://github.com/tosbaha)
 
 <!-- Your comment above here -->